requirements.txt
fastapi
uvicorn
scikit-learn
tensorflow
pillow
pandas
numpy
python-multipart